+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.5] - 2024-11-10
+- Improved handling of missing parent directory in writers/savers
+- Absorbed utility to drop None fields from into TOML writer 
+
 ## [0.2.4] - 2024-11-07
 - PyTorch model trainer now reports correct loss also for "sum" reduction
 - Added utility to drop None fields from dictionaries